Migrate to postgresql
Setup docker compose for postgresql, redis and rabbit use sqllite as
default local db and only use in memory for unit tests

resolves #160
resolves #159
resolves #158
