+++ src/WebApi/appsettings.json
@@ -1,23 +1,13 @@
 {  
   "ConnectionStrings": {
-    "mssql": "Server=(localdb)\\mssqllocaldb;Database=Exemplum;Trusted_Connection=True;MultipleActiveResultSets=true;"
+    "Exemplum": "Host=localhost;Username=postgres;Password=postgres;Database=Exemplum"
   },
   "WeatherForecast": {
     "BaseAddress": "http://api.openweathermap.org/"
   },
-  "UseInMemoryStorage": true,
-  "PublishIntegrationEvents": false,
-  "AllowedHosts": "*",
-  "HealthChecksUI": {
-    "HealthChecks": [
-      {
-        "Name": "Exemplum API",
-        "Uri": "https://localhost:5001/health/ready"
-      }
-    ],
-    "EvaluationTimeinSeconds": 10,
-    "MinimumSecondsBetweenFailureNotifications": 60
-  },
+  "UseInMemoryStorage": false, 
+  "WebAppUrl": "https://localhost:6001",
+  "AllowedHosts": "*",  
   "Auth0": {
     "Authority": "https://dev-ememplum.eu.auth0.com",
     "ApiIdentifier": "https://exemplumapi.com"
